style(plugin): style cleanups, add concern test
