+++ src/util/index.js
@@ -3,6 +3,8 @@ export { default as appendProps } from './appendProps'
 export { default as getClassDeclaration } from './getClassDeclaration'
 export { default as getExpressionIdentifier } from './getExpressionIdentifier'
 export { default as getEntryIdentifier } from './getEntryIdentifier'
+export { default as getFlowTypeKeys } from './getFlowTypeKeys'
+export { default as getTypeAliasIdentifier } from './getTypeAliasIdentifier'
 
 export { default as isReactComponent } from './isReactComponent'
 export { default as isReactImport } from './isReactImport'
