+++ test/fixtures/flow-non-concerning/actual.js
@@ -3,16 +3,16 @@ import React, { Component } from 'react';
 
 type Props = {
   name?: string;
-  _text?: string;
-  'aria-describedby'?: string;
-  children?: React.Node;
-};
+}
+
+type OtherProps = {
+  value?: string;
+}
 
 class Foo extends Component<Props> {
   render() {
     return <div {...this.props} />;
   }
-  static handledProps = ['_text', 'aria-describedby', 'children', 'name'];
 }
 
 export default Foo;
