+++ test/test.js
@@ -45,5 +45,7 @@ describe('fixtures', () => {
   fixtureAssert('stateless-arrow')
   fixtureAssert('stateless-assignment')
   fixtureAssert('stateless-predefined')
+
   fixtureAssert('flow')
+  fixtureAssert('flow-non-concerning')
 })
